fix: add PersonLink component and rewrite Routes in app
