+++ src/components/PeopleTable/index.ts
