+++ src/components/PersonLink/index.ts
@@ -1,0 +1,1 @@
+export * from './PersonLink';
